Extracted AxisGrid into shared folder
