+++ src/h5web/visualizations/heatmap/HeatmapVis.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react';
 import styles from './HeatmapVis.module.css';
 import ColorBar from './ColorBar';
 import { useHeatmapStyles, useProps, useValues } from './hooks';
-import AxisGrid from './AxisGrid';
+import AxisGrid from '../shared/AxisGrid';
 import Mesh from './Mesh';
 import Tooltip from './Tooltip';
 import HeatmapProvider from './HeatmapProvider';
@@ -12,6 +12,12 @@ import { useHeatmapConfig } from './config';
 function HeatmapVis(): JSX.Element {
   const props = useProps();
   const [mapAreaRef, heatmapStyles] = useHeatmapStyles();
+
+  const { dims, axisOffsets } = props;
+  const axisDomains = {
+    left: [0, dims[0]] as [number, number],
+    bottom: [0, dims[1]] as [number, number],
+  };
 
   const values = useValues();
   const initDataDomain = useHeatmapConfig(state => state.initDataDomain);
@@ -33,7 +39,7 @@ function HeatmapVis(): JSX.Element {
               <ambientLight />
               {/* Provide context again - https://github.com/react-spring/react-three-fiber/issues/262 */}
               <HeatmapProvider {...props}>
-                <AxisGrid {...props} />
+                <AxisGrid axisDomains={axisDomains} axisOffsets={axisOffsets} />
                 <Tooltip />
                 <Mesh />
               </HeatmapProvider>
