+++ src/h5web/visualizations/shared/IndexAxis.tsx
@@ -3,9 +3,9 @@ import { useMeasure } from 'react-use';
 import { AxisLeft, AxisBottom } from '@vx/axis';
 import { format } from 'd3-format';
 import { scaleLinear } from 'd3-scale';
-import styles from './HeatmapVis.module.css';
-import { adaptedNumTicks } from '../shared/utils';
-import { Domain } from '../shared/models';
+import styles from './AxisGrid.module.css';
+import { adaptedNumTicks } from './utils';
+import { Domain } from './models';
 
 type Orientation = 'bottom' | 'left';
